--- packages/client/tests/functional/adapter-errors/_matrix.ts
@@ -1,9 +1,0 @@
-import { defineMatrix } from '../_utils/defineMatrix'
-
-export default defineMatrix(() => [
-  [
-    {
-      provider: 'sqlite',
-    },
-  ],
-])
