--- packages/client/tests/functional/adapter-with-binary-engine/_matrix.ts
@@ -1,24 +1,0 @@
-import { defineMatrix } from '../_utils/defineMatrix'
-
-export default defineMatrix(() => [
-  [
-    {
-      provider: 'sqlite',
-    },
-    {
-      provider: 'postgresql',
-    },
-    {
-      provider: 'mysql',
-    },
-    {
-      provider: 'mongodb',
-    },
-    {
-      provider: 'cockroachdb',
-    },
-    {
-      provider: 'sqlserver',
-    },
-  ],
-])
