+++ packages/client/tests/functional/driver-adapters/binary-engine-error/_matrix.ts
@@ -1,0 +1,15 @@
+import { defineMatrix } from '../../_utils/defineMatrix'
+
+export default defineMatrix(() => [
+  [
+    {
+      provider: 'sqlite',
+    },
+    {
+      provider: 'postgresql',
+    },
+    {
+      provider: 'mysql',
+    },
+  ],
+])
