+++ packages/client/tests/functional/driver-adapters/binary-engine-error/prisma/_schema.ts
@@ -1,4 +1,4 @@
-import { foreignKeyForProvider, idForProvider } from '../../_utils/idForProvider'
+import { idForProvider } from '../../../_utils/idForProvider'
 import testMatrix from '../_matrix'
 
 export default testMatrix.setupSchema(({ provider }) => {
@@ -15,13 +15,6 @@ export default testMatrix.setupSchema(({ provider }) => {
   
   model User {
     id ${idForProvider(provider)}
-    profile Profile? 
-  }
-
-  model Profile {
-    id ${idForProvider(provider)}
-    userId ${foreignKeyForProvider(provider)} @unique
-    user User @relation(fields: [userId], references: [id])
   }
   `
 })
